pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyedgetwin"
version = "0.1.0"
description = "Python-first, container-native runtime for hybrid model deployment on edge devices"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Aero Shariati", email = "h.jannatabadi@iat.rwth-aachen.de"}
]
keywords = [
    "edge-computing",
    "digital-twin",
    "mqtt",
    "iot",
    "ml-inference",
    "hybrid-model",
    "industrial-iot",
    "time-series",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Hardware",
]
dependencies = [
    "paho-mqtt>=2.0.0,<3.0.0",
    "influxdb-client>=1.36.0,<2.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]
dashboard = [
    "streamlit>=1.28",
]

[project.urls]
Homepage = "https://github.com/aeroshariati/PyEdgeTwin"
Documentation = "https://github.com/aeroshariati/PyEdgeTwin#readme"
Repository = "https://github.com/aeroshariati/PyEdgeTwin.git"
Issues = "https://github.com/aeroshariati/PyEdgeTwin/issues"

[project.scripts]
pyedgetwin = "pyedgetwin.cli:main"

[project.entry-points."pyedgetwin.sinks"]
influxdb2 = "pyedgetwin.sinks.influxdb2:InfluxDB2Sink"
csv = "pyedgetwin.sinks.csv_sink:CSVSink"
stdout = "pyedgetwin.sinks.stdout:StdoutSink"

[project.entry-points."pyedgetwin.connectors"]
mqtt = "pyedgetwin.io.mqtt:MQTTConnector"

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.hatch.build.targets.wheel]
packages = ["src/pyedgetwin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["pyedgetwin"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "paho.*",
    "influxdb_client.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/pyedgetwin"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
